MLK-15959: ASoc: fsl: fix hifi4 driver build warning
When building sound/soc/fsl/fsl_hifi4.c file, a warning
occurs:
      warning: cast to pointer from integer of different
      size [-Wint-to-pointer-cast]
      (struct timestamp_info_t *)pext_msg->dtstamp;

By forced conversing int type to long type to fix this
issue.

+++ sound/soc/fsl/fsl_hifi4.c
@@ -1007,7 +1007,7 @@ int process_act_complete(struct fsl_hifi4 *hifi4_priv, u32 msg)
 			codec_iobuf_info->proc_id = pext_msg->proc_id;
 			hifi4_priv->is_done = 1;
 			hifi4_priv->dpu_tstamp =
-			    (struct timestamp_info_t *)pext_msg->dtstamp;
+			  (struct timestamp_info_t *)((long)pext_msg->dtstamp);
 			hifi4_priv->ret_status = pext_msg->ret;
 			complete(&hifi4_priv->cmd_complete);
 		}
